test: update role-related tests for clarity and functionality, including method renaming and parameter adjustments
