+++ api/app/src/test/kotlin/packit/integration/controllers/RoleControllerTest.kt
@@ -2,7 +2,6 @@ package packit.integration.controllers
 
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpMethod
@@ -12,7 +11,6 @@ import packit.integration.IntegrationTest
 import packit.integration.WithAuthenticatedUser
 import packit.model.*
 import packit.model.dto.*
-import packit.repository.PacketGroupRepository
 import packit.repository.PermissionRepository
 import packit.repository.RoleRepository
 import packit.repository.UserRepository
@@ -36,19 +34,6 @@ class RoleControllerTest : IntegrationTest()
 
     @Autowired
     private lateinit var roleService: RoleService
-
-    @Autowired
-    private lateinit var packetGroupRepository: PacketGroupRepository
-
-    private var packetGroup = PacketGroup(
-        name = "test-name",
-    )
-
-    @AfterEach
-    fun cleanupData()
-    {
-        packetGroupRepository.delete(packetGroup)
-    }
 
     private val testCreateRole = CreateRole(
         name = "testRole",
@@ -388,86 +373,5 @@ class RoleControllerTest : IntegrationTest()
             jacksonObjectMapper().readTree(result.body).get("users").first().get("username").asText()
         )
         assertEquals(1, roleRepository.findByName(testRole.name)!!.users.size)
-    }
-
-    @Test
-    @WithAuthenticatedUser(authorities = ["packet.manage:packet:test:123"])
-    fun `user without permission cannot update read permission on roles`()
-    {
-        val updatePacketReadRoles = jacksonObjectMapper().writeValueAsString(
-            UpdatePacketReadRoles(
-                packetGroupName = "test-name",
-                roleNamesToAdd = setOf(),
-                roleNamesToRemove = setOf()
-            )
-        )
-        val result = restTemplate.exchange(
-            "/roles/read-permissions",
-            HttpMethod.PUT,
-            getTokenizedHttpEntity(data = updatePacketReadRoles),
-            String::class.java
-        )
-
-        assertEquals(result.statusCode, HttpStatus.UNAUTHORIZED)
-    }
-
-    @Test
-    @WithAuthenticatedUser(authorities = ["packet.manage"])
-    fun `user with packet manage can update read permission on roles`()
-    {
-        packetGroup = packetGroupRepository.save(
-            packetGroup
-        )
-        val roleNamesToAdd = setOf("testRole1", "testRole2")
-        val roleNamesToRemove = setOf("testRole3", "testRole4")
-        roleRepository.saveAll(
-            roleNamesToAdd.map {
-                Role(name = it)
-            }
-        )
-        val rolesToRemove = roleRepository.saveAll(
-            roleNamesToRemove.map {
-                Role(name = it)
-            }
-        ).onEach {
-            it.rolePermissions = mutableListOf(
-                RolePermission(
-                    role = it,
-                    permission = permissionsRepository.findByName("packet.read")!!,
-                    packetGroup = packetGroup
-                )
-            )
-        }
-        roleRepository.saveAll(rolesToRemove)
-        val updatePacketReadRoles = jacksonObjectMapper().writeValueAsString(
-            UpdatePacketReadRoles(
-                packetGroupName = packetGroup.name,
-                roleNamesToAdd = roleNamesToAdd,
-                roleNamesToRemove = roleNamesToRemove
-            )
-        )
-
-        val result = restTemplate.exchange(
-            "/roles/read-permissions",
-            HttpMethod.PUT,
-            getTokenizedHttpEntity(data = updatePacketReadRoles),
-            String::class.java
-        )
-
-        assertEquals(result.statusCode, HttpStatus.NO_CONTENT)
-        val addedRoles = roleRepository.findByNameIn(roleNamesToAdd.toList())
-        val removedRoles = roleRepository.findByNameIn(roleNamesToRemove.toList())
-        removedRoles.forEach {
-            assertEquals(
-                it.rolePermissions.size,
-                0
-            )
-        }
-        addedRoles.forEach {
-            assertEquals(
-                it.rolePermissions.first().packetGroup?.id,
-                packetGroup.id
-            )
-        }
     }
 }
